# Provide Hash#count for performance improvement [Feature #17758]