@marcandre 2021-03-28
Best to mention this change in the NEWS file, like we did for `Array#max` and `Array#min` [in 2.4.0](https://github.com/ruby/ruby/blob/master/doc/NEWS-2.4.0)