@kachick 2021-03-31
Thanks for your review! 🙏 

>Oh, I notice that you call super. This is unusual, once we enter the C world we normally call the implementation directly...

😮 Oh... Using `super` in C is a bad way?

When I implemented this, I referenced `Range#count`.

https://github.com/ruby/ruby/blob/66d2fc7989d741bf5a73286233139901cecb4fc2/range.c#L1714-L1748

It looks to use `super` 🤔 

And I thought calling `super` method might keep compatibility with old rubies behaviors as possible...